requirements.txt
PySide6
pandas
pyarrow
Pillow
numpy
matplotlib
